界面修改完成

+++ StudyContribution/ClassInfoViewController.m
@@ -105,7 +105,7 @@
       */
  - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
  {
-         NSLog(@"numberOfRowsInSection %d", section);
+         NSLog(@"numberOfRowsInSection %ld", (long)section);
          if (0 == section) {
                  // 第0组有多少行
                  return 4;
@@ -120,7 +120,7 @@
       */
  - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
  {
-     NSLog(@"cellForRowAtIndexPath %d %d", indexPath.section, indexPath.row);
+     NSLog(@"cellForRowAtIndexPath %ld %ld", (long)indexPath.section, (long)indexPath.row);
     
      //    indexPath.section; // 第几组
      //    indexPath.row; // 第几行
